+++ src/javascripts/events/guestNavEvents.js
@@ -1,5 +1,5 @@
-import showReservations from '../components/reservations';
-import getReservations from '../helpers/data/reservationData';
+import { showReservations } from '../reservations/reservations';
+import getReservations from '../reservations/reservationData';
 import getMenuItems from '../menu items/menuData';
 import showMenuItems from '../menu items/menu';
 import showStaff from '../components/showStaff';
